+++ RechargeSharp.v2021-11/Endpoints/Subscriptions/SubscriptionService.Endpoints.cs
@@ -7,14 +7,14 @@ namespace RechargeSharp.v2021_11.Endpoints.Subscriptions;
 public interface ISubscriptionService
 {
     Task<SubscriptionService.CreateSubscriptionTypes.Response> CreateSubscriptionAsync(SubscriptionService.CreateSubscriptionTypes.Request request);
-    Task<SubscriptionService.GetSubscriptionTypes.Response?> GetSubscriptionAsync(int subscriptionId);
-    Task<SubscriptionService.UpdateSubscriptionTypes.Response> UpdateSubscriptionAsync(int subscriptionId, SubscriptionService.UpdateSubscriptionTypes.Request request);
-    Task DeleteSubscriptionAsync(int subscriptionId);
+    Task<SubscriptionService.GetSubscriptionTypes.Response?> GetSubscriptionAsync(long subscriptionId);
+    Task<SubscriptionService.UpdateSubscriptionTypes.Response> UpdateSubscriptionAsync(long subscriptionId, SubscriptionService.UpdateSubscriptionTypes.Request request);
+    Task DeleteSubscriptionAsync(long subscriptionId);
     Task<SubscriptionService.ListSubscriptionsTypes.Response> ListSubscriptionsAsync(SubscriptionService.ListSubscriptionsTypes.Request? request);
-    Task<SubscriptionService.ChangeSubscriptionsNextChargeDateTypes.Response> ChangeSubscriptionsNextChargeDateAsync(int subscriptionId, SubscriptionService.ChangeSubscriptionsNextChargeDateTypes.Request request);
-    Task<SubscriptionService.ChangeSubscriptionsAddressTypes.Response> ChangeSubscriptionsAddressAsync(int subscriptionId, SubscriptionService.ChangeSubscriptionsAddressTypes.Request request);
-    Task<SubscriptionService.CancelSubscriptionTypes.Response> CancelSubscriptionAsync(int subscriptionId, SubscriptionService.CancelSubscriptionTypes.Request request);
-    Task<SubscriptionService.ActivateSubscriptionTypes.Response> ActivateSubscriptionAsync(int subscriptionId);
+    Task<SubscriptionService.ChangeSubscriptionsNextChargeDateTypes.Response> ChangeSubscriptionsNextChargeDateAsync(long subscriptionId, SubscriptionService.ChangeSubscriptionsNextChargeDateTypes.Request request);
+    Task<SubscriptionService.ChangeSubscriptionsAddressTypes.Response> ChangeSubscriptionsAddressAsync(long subscriptionId, SubscriptionService.ChangeSubscriptionsAddressTypes.Request request);
+    Task<SubscriptionService.CancelSubscriptionTypes.Response> CancelSubscriptionAsync(long subscriptionId, SubscriptionService.CancelSubscriptionTypes.Request request);
+    Task<SubscriptionService.ActivateSubscriptionTypes.Response> ActivateSubscriptionAsync(long subscriptionId);
 }
 
 public partial class SubscriptionService : ISubscriptionService
@@ -33,21 +33,22 @@ public partial class SubscriptionService : ISubscriptionService
         return responseJson;
     }
 
-    public async Task<GetSubscriptionTypes.Response?> GetSubscriptionAsync(int subscriptionId)
+    public async Task<GetSubscriptionTypes.Response?> GetSubscriptionAsync(long subscriptionId)
     {
         var requestUri = $"/subscriptions/{subscriptionId}";
         var responseJson = await _rechargeApiCaller.GetNullableAsync<GetSubscriptionTypes.Response> (requestUri);
         return responseJson;
     }
     
-    public async Task<UpdateSubscriptionTypes.Response> UpdateSubscriptionAsync(int subscriptionId, UpdateSubscriptionTypes.Request request)
+    public async Task<UpdateSubscriptionTypes.Response> UpdateSubscriptionAsync(long subscriptionId,
+        UpdateSubscriptionTypes.Request request)
     {
         var requestUri = $"/subscriptions/{subscriptionId}";
         var responseJson = await _rechargeApiCaller.PutAsync<UpdateSubscriptionTypes.Request, UpdateSubscriptionTypes.Response>(request,requestUri);
         return responseJson;
     }
     
-    public async Task DeleteSubscriptionAsync(int subscriptionId)
+    public async Task DeleteSubscriptionAsync(long subscriptionId)
     {
         var requestUri = $"/subscriptions/{subscriptionId}";
         await _rechargeApiCaller.DeleteAsync(requestUri);
@@ -61,28 +62,31 @@ public partial class SubscriptionService : ISubscriptionService
         return responseJson;
     }
     
-    public async Task<ChangeSubscriptionsNextChargeDateTypes.Response> ChangeSubscriptionsNextChargeDateAsync(int subscriptionId, ChangeSubscriptionsNextChargeDateTypes.Request request)
+    public async Task<ChangeSubscriptionsNextChargeDateTypes.Response> ChangeSubscriptionsNextChargeDateAsync(
+        long subscriptionId, ChangeSubscriptionsNextChargeDateTypes.Request request)
     {
         var requestUri = $"/subscriptions/{subscriptionId}/set_next_charge_date";
         var responseJson = await _rechargeApiCaller.PostAsync<ChangeSubscriptionsNextChargeDateTypes.Request, ChangeSubscriptionsNextChargeDateTypes.Response>(request, requestUri);
         return responseJson;
     }
     
-    public async Task<ChangeSubscriptionsAddressTypes.Response> ChangeSubscriptionsAddressAsync(int subscriptionId, ChangeSubscriptionsAddressTypes.Request request)
+    public async Task<ChangeSubscriptionsAddressTypes.Response> ChangeSubscriptionsAddressAsync(long subscriptionId,
+        ChangeSubscriptionsAddressTypes.Request request)
     {
         var requestUri = $"/subscriptions/{subscriptionId}/change_address";
         var responseJson = await _rechargeApiCaller.PostAsync<ChangeSubscriptionsAddressTypes.Request, ChangeSubscriptionsAddressTypes.Response>(request, requestUri);
         return responseJson;
     }
     
-    public async Task<CancelSubscriptionTypes.Response> CancelSubscriptionAsync(int subscriptionId, CancelSubscriptionTypes.Request request)
+    public async Task<CancelSubscriptionTypes.Response> CancelSubscriptionAsync(long subscriptionId,
+        CancelSubscriptionTypes.Request request)
     {
         var requestUri = $"/subscriptions/{subscriptionId}/cancel";
         var responseJson = await _rechargeApiCaller.PostAsync<CancelSubscriptionTypes.Request, CancelSubscriptionTypes.Response>(request, requestUri);
         return responseJson;
     }
     
-    public async Task<ActivateSubscriptionTypes.Response> ActivateSubscriptionAsync(int subscriptionId)
+    public async Task<ActivateSubscriptionTypes.Response> ActivateSubscriptionAsync(long subscriptionId)
     {
         var requestUri = $"/subscriptions/{subscriptionId}/activate";
         var responseJson = await _rechargeApiCaller.PostAsync<ActivateSubscriptionTypes.Response>(requestUri);
